created separate app for dashboard
