+++ dashboard/templates/dashboard/three-way-toggle.html
