+++ dashboard/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+app_name = "dashboard"
+urlpatterns = [
+    path("", views.DashboardView.as_view(), name="dashboard"),
+]
