+++ lymph_interface/loggers.py
@@ -1,6 +1,4 @@
 import logging
-import time
-from typing import Dict, Any
 
 
 class ModeLoggerMixin(object):
