+++ lymph_interface/templates/index.html
@@ -44,7 +44,7 @@
         <h1 class="title">Features</h1>
         <div class="tile is-ancestor">
             <div class="tile is-parent is-vertical">
-                <a href="{% url 'patients:dashboard' %}" class="tile is-child notification is-primary is-light">
+                <a href="{% url 'dashboard:dashboard' %}" class="tile is-child notification is-primary is-light">
                     <div class="media">
                         <figure class="media-left">
                             <span class="icon is-large">
