+++ lymph_interface/templates/navbar.html
@@ -24,7 +24,7 @@
                     Download
                 </a>
 
-                <a href="{% url 'patients:dashboard' %}" class="navbar-item" title="explore patterns of progression">
+                <a href="{% url 'dashboard:dashboard' %}" class="navbar-item" title="explore patterns of progression">
                     Dashboard
                 </a>
             </div>
