+++ lymph_interface/urls.py
@@ -24,6 +24,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('accounts.urls')),
     path("patients/", include("patients.urls")),
+    path("dashboard/", include("dashboard.urls")),
     path("", views.index, name="index"),
 ]
