+++ patients/ioports.py
@@ -1,11 +1,9 @@
 from django.db import IntegrityError
-from django.forms import ValidationError
 from django.db.models import QuerySet
 
 import numpy as np
 import pandas as pd
-import dateutil.parser as dp
-from typing import List, Union, Optional, Dict, Any
+from typing import List
 import logging
 logger = logging.getLogger(__name__)
