--- patients/static/patients/dashboard.css
@@ -1,83 +1,0 @@
-body {
-    background-color: #dbe6eb;
-}
-
-input[type="checkbox"].checkbox:checked + label {
-    background-color: #005EA8;
-    color: white;
-}
-input[type="checkbox"].checkbox:checked:hover + label {
-    background-color: #005392;
-    color: white;
-}
-
-input[type="radio"].radio:checked + label {
-    background-color: #005EA8;
-    color: white;
-}
-
-.tumor.button {
-    min-width: 200px;
-}
-.stats.button {
-    min-width: 60px;
-}
-
-input[type="radio"].radio:checked + label.tumor {
-    background-color: #005EA8;
-    color: white;
-}
-input[type="radio"].radio:checked:hover + label.tumor {
-    background-color: #005392;
-    color: white;
-}
-
-div.tag.pos {
-    min-width: none;
-    padding: 0;
-    background-color: #AE0060;
-}
-div.tag.neg {
-    min-width: none;
-    padding: 0;
-    background-color: #00AFA5;
-}
-div.tag.unk {
-    min-width: none;
-    padding: 0;
-    background-color: #C5D5DB;
-}
-
-div.tag.stat-tag {
-    width: 34px;
-}
-
-.barplot {
-    background-image:
-        linear-gradient(#AE0060,#AE0060),
-        linear-gradient(#BDCFD6,#BDCFD6),
-        linear-gradient(#00AFA5,#00AFA5);
-    background-repeat:no-repeat;
-}
-
-.barplot-legend-left {
-    text-align: left;
-}
-.barplot-legend-center {
-    text-align: center;
-}
-.barplot-legend-right {
-    text-align: right;
-}
-
-.barplot-legend {
-    width: 40px;
-}
-
-.barplot-tablecell {
-    width: 100%;
-}
-
-span[data-tooltip] {
-    border-bottom: none;
-}
